pyproject.toml
[project]
name = "course-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
